Moved random functions from util.hpp to util_random.hpp
